Move loss to new module
