+++ src/invrs_gym/loss/__init__.py
@@ -1,0 +1,1 @@
+from invrs_gym.loss import transmission_loss as transmission_loss
